Improve HTMLForm's support for adding content in and around the form, to improve versatility.  Groundwork for being able to validate, filter and submit the CreateAccount input via HTMLForm::trySubmit(), which will give extensions huge flexibility to play with the form input.
Also change the wrapping on both forms to use a <fieldset>, which is visually more consistent with the other forms we have floating around the site.  Add visualClear to all HTMLForm forms, and remove all the redundant CSS for the login form, quite a bit of which has been dead for a while.

git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk/phase3@56896 dd0e9695-b195-4be7-bd10-2dea1a65a6b6