Reverted r59548 (reopens bug 21086).
wfWikiID() is used as a way to identify and connect to a given database in LoadBalancer and various extensions, that's why it must contain the database name and table prefix. Perhaps its use as a public identifier should be optional, but in that case the public identifier needs to be customised, not the return value of wfWikiID().



git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk/phase3@60960 dd0e9695-b195-4be7-bd10-2dea1a65a6b6